requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
pydantic==2.6.2
pydantic_core==2.16.3
google-api-python-client==2.125.0
openai==1.12.0
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
google-cloud-secret-manager==2.19.0